this starts to fucking annoy

+++ kernel/tr/executor/base/PPBase.cpp
@@ -58,25 +58,25 @@ const char* SednaXQueryException::createMessage(char* buffer) const
         user_error_code_entries[internal_code].code,
         user_error_code_entries[internal_code].descr);
 
-    CHECK_SNPRINTF_RET(len, buffer, buf0);
+    CHECK_SNPRINTF_RET((int)len, buffer, buf0);
 
     if (err_msg[0] != 0) {
         len = snprintf(buffer, EXCEPTION_BUFFER_SIZE - (buffer - buf0),
             "Details: %s\n", err_msg);
-        CHECK_SNPRINTF_RET(len, buffer, buf0);
+        CHECK_SNPRINTF_RET((int)len, buffer, buf0);
     }
 
     if (xquery_line != 0)
     {
         len = snprintf(buffer, EXCEPTION_BUFFER_SIZE - (buffer - buf0),
             "Query line: %d, column: %d \n", xquery_line, xquery_col);
-        CHECK_SNPRINTF_RET(len, buffer, buf0);
+        CHECK_SNPRINTF_RET((int)len, buffer, buf0);
     }
 
 #if (EL_DEBUG == 1)
     len = snprintf(buffer, EXCEPTION_BUFFER_SIZE - (buffer - buf0),
         "Position: [%s, %s, %d]\n", file, function, line);
-    CHECK_SNPRINTF_RET(len, buffer, buf0);
+    CHECK_SNPRINTF_RET((int)len, buffer, buf0);
 #endif
     return buf0;
 }
